python/深度学习/图片识别-efficientb4/requirements.txt
flyai
numpy
Pillow
pytorch==1.1.0
opencv-python